refactor(HotBoardSelectionStrategy): 전략 패턴에서 저장과 조회 로직을 분리
